# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing tasks and sandboxes.",
        "title": "Platform Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/tasks": {
            "get": {
                "description": "Returns task summaries ordered by most recently updated.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "List tasks for the authenticated user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.taskListItem"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new task. Optional git_url clones the repository at provision time. Private repos (git@ or ssh://) require an SSH key to be configured.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Create a task",
                "parameters": [
                    {
                        "description": "Create task request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.createTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_api.createTaskResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request body or missing SSH key for private repo",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    },
                    "500": {
                        "description": "failed to create task",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "description": "Retrieve task state by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.getTaskResponse"
                        }
                    },
                    "404": {
                        "description": "task not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to get task",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a task and its associated sandbox.",
                "tags": [
                    "tasks"
                ],
                "summary": "Delete a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "task not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to delete task",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/tasks/{id}/history": {
            "get": {
                "description": "Retrieve the conversation history for a task. Requires fileStore to be configured.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get task conversation history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Raw session entries",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "404": {
                        "description": "task not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to get history",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "history storage not configured",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/tasks/{id}/messages": {
            "post": {
                "description": "Lazily provisions the task sandbox on first use and streams the assistant response back to the caller. Accepts JSON or multipart/form-data (with image attachments).",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Send a message to a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Send message request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.sendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Streamed assistant response",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "prompt is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "task not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to get task",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "failed to provision sandbox",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/tasks/{id}/permissions": {
            "post": {
                "description": "Approve or deny a canUseTool permission request that has paused the agent session.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Respond to a pending tool permission request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permission decision",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.respondToPermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "task not found or no pending permission",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "failed to respond to permission",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/tasks/{id}/questions": {
            "post": {
                "description": "Submit answers to a clarifying question that has paused the agent session.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Respond to a pending AskUserQuestion request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Question answers",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.respondToQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "task not found or no pending question",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "failed to respond to question",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.createTaskRequest": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "env": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "git_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_api.createTaskResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "internal_api.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.respondToPermissionRequest": {
            "type": "object",
            "required": [
                "decision"
            ],
            "properties": {
                "decision": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.sendMessageRequest": {
            "type": "object",
            "required": [
                "prompt"
            ],
            "properties": {
                "permissionMode": {
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.taskListItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "error_msg": {
                    "type": "string"
                },
                "git_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}